// src/lib/miden/db/transaction-types.ts
export enum ITransitionStatus {
  Queued,
  Generating,
  Completed,
  Failed
}

export interface ITransition {
  id: string;
  transitionId?: string;
  transactionDbId: string;
  index: number;
  inputRecordIds: string[];
  outputRecordIds: string[];
  address: string;
  chainId: string;
  program: string;
  functionName: string;
  status: ITransitionStatus;
  inputsJson: string;
  outputsJson?: string;
  initiatedAt: number;
  completedAt?: number;
  json?: string;
  isFee: number;
}